requirements.txt
torch
torchvision
scikit-learn
numpy
pillow
joblib
h5py
matplotlib
pylatexenc
qiskit
qiskit_algorithms
qiskit_machine_learning
scikit-image
pandas
astroquery
datasets